+++ apps/cli/cli_pipe.c
@@ -318,7 +318,7 @@ pipe_showas_fn(clixon_handle h,
             goto done;
         break;
     case FORMAT_JSON:
-        if (clixon_json2file(stdout, xt, pretty, cligen_output, 1, 0) < 0)
+        if (clixon_json2file(stdout, xt, pretty, cligen_output, 1, 0, 0) < 0)
             goto done;
         break;
     case FORMAT_TEXT: